WikibaseRepo: add means to get entity on server
Done through mwbot
This makes our tests (ts-jest) use the tsconfig.all.json until we get to
re-organize the test folder to be able to test client-only and
server-only (and possibly mixed) individually. This is to ensure the
typeRoots defined (for server) are not made obsolete by being excluded
thereafter.

Bug: T207467