production build: ensure 404s are reported correctly
In production builds, with minification enabled, exceptions that were
created inside the server bundle could be neither analyzed by class (as
classes loaded inside and outside the bundle are effectively seens as
two independent ones) nor class name (as minification caused the name to
be shortened).
This introduces a dedicated error response object for this rare scenario
for the server to be able to further analyze the reason of errors from
within the bundle, e.g. to tell apart a server error from an entity that
can not be found.
This also disables minifcation in the node environment of which there is
no clear advantage (size is not a factor) but which deprives us of the
ability to identify the error response by name.

Bug: T211573