env: explicitly set variables relevant for server process
This resolves a test todo (mocking based on .env) and allows us to
specify variables relevant to the runtime on server start instead of on
server/[blubber] image build.
This also introduces the concept of services that are reusable accross
multiple requests (e.g. mwbot) and provides a way of passing it to the
individual process handler.

Bug: T211675