tslib: ensure presence in prod installation
We run typescript with the configuration `importHelpers: true`,
consequently the compiler will inject helper functions into the
generated output that are called at run-time. These helper functions
reside in the tslib library which consequently has to be installed.
So far it has only been installed as a 3rd party dependency of some of
our dev dependencies. Consequently running the server with only prod
dependencies installed, as we plan for production, failed.
See
https://blog.mariusschulz.com/2016/12/16/typescript-2-1-external-helpers-library
https://www.npmjs.com/package/tslib

Bug: T211675