blubber: verifiy checked in dist folder is clean
Verify that what is checked in as dist folder (containing the JS, CSS
to be used on the client side) matches the build result - i.e. the dev
last changing the source files did not forget to check in the lastest
build result.
In contrast to what happenend before, the server build is now executed
(by blubber, when the image is built, as root) _before_ the moment when
the client code is build (during test).
The presence of the cache folder from the server build
changes the client build outcome (something we learned earlier),
and removing it requires ownership of the files, so
we clean the cache after the server build polluted it.
It is a bit unfortunate that the two topics got intertwined by deleting
the cache during the client build became impossible as it was written by
the root user during blubber image build.

Bug: T214141
Change-Id: I7a8711c2dcde31c5f0900baa872e465ba02aabee