SCSS: wrap all selectors to overcome reset CSS
Allows the individual components not having to use an extra class
(otherwise unrelated to them) to build a strong selector chain in order to
overcome the reset CSS used in App.vue.
Consequently the indivdual components have no more technical reason to
refer to classes defined outside of them and this practice is to be
avoided to maintain the benefits of isoliation (see BEM); an ADR/doc
should follow.

Change-Id: I38b890e7fc3ec34664d1d79281db700a476d7a8f