npm: separate build and runtime dependencies
While all of the packages mentioned in package.json are dependencies,
those in devDependencies are not needed to "run" the application but
only to create the artifacts required to do so.
Following this idea, the (non dev) dependencies should only contain
packages that will be used during application run time (i.e. SSR).

Change-Id: Ief4c009b9f6ff89b4dbcb48485ad83861e2f95e7