npm: check in linux package-lock
The package-lock.json contained optional packages that resulted from an
install on a mac. We can only have one or the other so we go with the
majority. A ticket about this is open for quite some time already but no
luck so far: https://github.com/npm/npm/issues/17722

Change-Id: I3df9291eef24d39fa141a599f5211323e50f0370