QueryValidator test: fail for the right reasons
Ensure that failing tests do indeed fail because of the intended
reason and not as side effect of other changes.
Inspecting the errors was one idea that came to mind and also results
in a rather self-documenting test.
It is still quite repetitive, property-based testing comes to mind
but we do not really want/need to (re)test our validation library that
heavily.

Change-Id: I3ff9945e0d29731b5932d27e175509bf8cb96df4