beta hosting: add ansible playbook
Adds a basic playbook to setup the system we run for serving SSR
to beta.
The configuration (as interpolated from the template and the only
set of vars) is the one currently in place; i.e. I applied that.

This completely lacks the part setting up the regularly rebuilding the
docker image that is used by the service.

Is a follow-up to Iac55d84d8d1f287901f8c82513f9b9578fb399bb - had the
infrastructure files existed at the time it would have changed together
with the other files.

Bug: T210615
Bug: T212196
Change-Id: Ia2418897961cbe89ae6cea3c9fbe1380e9ec9bfb