shrinking-flex-element: add mixin
Create mixin to centralize documentation for this workaround and wrap in
a catchy title so purpose becomes obvious on the using side.
Follow-up to I81d0ad53fdab117b4b045cf694a1b61f5c80f456

This also moves one of the uses of this to where it logcially belongs
(took quite a bit of forensics to make sense of it), declaring the
primary Fingerprint - the content of Termbox - as a shrinking flex
element instead of all fingerprints (even when potentially not used
in a flex container).
In order to select this element it uses the "*-inner" pattern
known from Fingerprint - definitely not impressive but consistent.

Bug: T217244
Change-Id: I5313709512c782e155095aa6aefbeef03cd52306