npm: clean package-lock
Ensure package-lock.json is clean after install.
Unlike with I3df9291eef24d39fa141a599f5211323e50f0370 we are not quite
sure which factors led to the file checked in as it was. Word is this
was done on a Kubuntu with npm 6.4.1

For reasons not clear to me it makes a difference if you remove your
node_modules folder beforehand. Make sure to do so once to not have a
diff (now, and hopefully in the future).

tarrow pointed to https://stackoverflow.com/q/45022048 which indicates
that only ever using `npm ci` could be an answer to avoid any of this
locally, but as we were doing fine with `npm i` so far I think we should
at least try to get it back into shape.

Change-Id: I3ca9929129bb01ebc871003667a57418eff10b7b