wikimedia-ui-base: add and allow variable use
Adds https://www.npmjs.com/package/wikimedia-ui-base and a way to allow
access to its (LESS & CSS) variables in SCSS.
The resulting SCSS file is - controversially - stored alongside the
wikimedia-ui-base project in node_modules to fight temptation to edit
it directly.

It has to be noted that this is not feature complete as it uses CSS
custom properties as the base of the translation, and as they can depend
on other properties, which is not translated properly, some variables
still contain "var()" expressions that SCSS does not understand.

It is imported into our existing variables file where individual values
can be overwritten/added.

This replaces two, so far hard coded, values on our side with their
representation in wikimedia-ui-base as a PoC.
As a next step, now that we have access to the "official" variables, we
should also rethink how we use them inside our project.

Change-Id: I9de8420b2205442ac448a634af25fe75d7b3ea5c