mockup-entry: add origin param to dev API calls
Ensure that client-side API calls are performed with the appropriate
origin value so that the (dev) mediawiki has a chance to answer them and
the requests are not prevented by your friendly browser.
See https://www.mediawiki.org/wiki/Manual:CORS
and set your wiki's $wgCrossSiteAJAXdomains to include the protocol,
host and port (e.g. 'http://localhost:8080') of your termbox dev mockup.
This works by adding a request interceptor that is not required in
production because there the Javascript will be served by the same
system (through ResourceLoader) that the rest of the page lives, too.

Bug: T218577
Change-Id: I8f43486e5bb07315135aeb3a9bfcc71959f34d00