wikimedia-ui-base: convert all variables to SCSS
Build an abstract syntax tree from the less variables file, replace
variables referencing other variables with a copy of the value
originally used to define it, then convert it to scss.
The resulting variables file will not contain self-references , i.e.
possible overrides (unlikely as we want to inherit the values) would
have to happen on a per-variable basis.

This is a followup to overcome the shortcomings of
I9de8420b2205442ac448a634af25fe75d7b3ea5c

Bug: T220363
Change-Id: I2abb5332bef568765f6ed62bf00cebb50fa7d94f