Bug Fix for Array is not extensible when adding keys to state objects
Vuex states have something similar to Object.preventExtensions() [1]

Vue's reactivity requires us to be careful when mutating states
so that components observing the state can update automatically. [2]

This patch changes all mutations to replace objects in states
so that in the event we add a new key everything works as planned.
It does this using the "spread" style and Vue.set from the Vuex docs for sepcific cases. [2]

Let's emphasize following rule:
 For adding a single pair we are using Vue.set, for entire
 sets/lists/arrays/objects we use spread.

[1] https://developer.mozilla.org/de/docs/Web/JavaScript/Reference/Fehler/Cant_define_property_object_not_extensible
[2] https://vuex.vuejs.org/guide/mutations.html#mutations-follow-vue-s-reactivity-rules

Bug: T220379
Change-Id: I78debaa1518bc29557a13b5addc9e02d30e8acc8