Use TermTextField to edit aliases individually
Creates individual TermTextFields for each alias, also creates a new
empty TermTextField on mount and creates more empty TermTextFields when
the final TermTextField has input.

This patch has a local copy of the aliases passed from the parent
component and thus is not (yet) bound to the store. We included the
blank alias in the main list of local aliases rather than having a
special permanently blank TermTextField because jumping the focus
between the blank and the "new" alias was buggy on mobile and deemed
inelegant.

Bug: T220392
Bug: T220384

Change-Id: I3af3a13606172a8bd2b13e6d6a752fac7d030648