Start wiring up the store to AliasesEdit
Use new actions to edit aliases and remove deleted aliases from the
store. Aliases are edited on TermTextField input and removed if needed
on blur.

In the AliasesEdit tests we simply look for actions related to editing
and deleting to be correctly fired and do not assert that the complete
feedback loop propagates through the parent via the props.

Bug: T220394
Co-Authored-By: Jakob <jakob.warkotsch@wikimedia.de>
Change-Id: I2f320e3e1a6f2f912d813e6a76f9a702dbc41cc5