MonolingualFingerprintView: grow despite little content
Allow the MonolingualFingerprintView to grow within the limits of
max-width even if its child elements are not filled with content to the
extend that this happens "by itself".
Without this, in edit mode, input elements - which through their border
reveal their true width - would only be about as wide as their text
content. Could be observed e.g. in mock entry point for English as
primary language, which has a very short label.

Bug: T218690 (discovered during)
Change-Id: Iacd4f405d258e2013a0987829c23cded86ab8f80