Move Docker image build steps after configuration step.
Without configuring a .env files with proper values, building
the image is not possible or will result in wrong image being built
if the user just pust any values in .env to by-pass that step.

Moving it after configuration step makes sure the user will know
what they are doing when preparing .env and will build the right
image for their needs.

Change-Id: I430a34458f14536f12b4b0ecba81646ebe2ab369