Remove repoUrl from base url
In order to prevent us from accidentally making requests to other
sites and throwing CORS errors we can always make them relative.

This mimics the default action of MobileFrontend [1].

This patch was selected in preference to getting wbRepo.url to reflect
the host to which the client is connecting because Mediawiki doesn't
actually know this. See for example: mw.config.get('wgServer') on m.wikidata.org
is equal to `//www.wikidata.org`.

The CSR dev mode was fixed by using the vue dev server proxy. [2]

[1] https://phabricator.wikimedia.org/diffusion/EMFR/browse/master/README.md$220
[2] https://cli.vuejs.org/config/#devserver-proxy

Bug: T220995
Change-Id: Id0af88b03354e242c3b9db1aae18d91379d9e3d0