Add x-monitor and x-amples
Only adding checking to _info due to lack of clarity about when
and how these examples are tested.

The output (and status code etc.)
of the /termbox endpoint is not just a function of the query parameters
but also of the response of a remote service. In this case the validity
of the query parameters depends upon the state of that remote service.

Bug: T222358
Change-Id: I520f90acbbd1973780101edd102c91a3fa3e633b