Change EventEmittingButton for mw-styles
This patch makes several changes to get the possibility to create
buttons according to ooui style guide [1]. This is not a complete set of
all possible buttons. It contains: normal, primary-progressive and
frameless-progressive.
To test this at current state of termbox use one of existing buttons and
change them into 'normal', 'primaryProgressive' or
'framelessProgressive'.

[1]: https://doc.wikimedia.org/oojs-ui/master/demos/?page=widgets&theme=wikimediaui&direction=ltr&platform=desktop#demo-section-buttons

Bug: T222695
Change-Id: I5e71387e49ad9b7f4d04abedf6fb91b704b0a7e9