server: config validation improvements + logging
Move config validation to dedicated file, base on a light specification,
coerce integer types.
I looked into a couple of libraries but they all were quite bulky, most
a little off-topic, few added type safety - this covers our use case and
is not too excessive.
This also reduces the logger to what the one injected in production is
capable of and wires it up. Log levels were improvised based on [0]&[1].

[0] https://github.com/wikimedia/service-runner#features
[1] https://www.mediawiki.org/wiki/ServiceTemplateNode/CodingGuide#Logging

Bug: T222736
Change-Id: I7a00a3d8a2e53063da9126166335a194b6a1f3e4