dev: align image used with prod/CI
Align the image we use for development with docker-compose with the one
used in CI.
Pipe dream: base the dev image on an infrastructure image created by
blubber to avoid diverging. Created [0] and a bucket list entry to be
able to do that without making blubber itself a system requirement for
our development tool chain.

When switching to this, make sure to remove your pre-existing
node_modules folder and run the steps from "Building Docker image" from
the README.

[0] https://wikitech.wikimedia.org/wiki/Talk:Blubber/Download#Docker_image

Change-Id: I2617592c42edf07bd1f6da774324c5da5fca2689