EventEmittingButton: re-establish rtl support
This restores RTL support. Due to the abysmal browser support for the
:dir() CSS pseudo-class we need to reference an actual attribute.
To avoid passing around and attaching attributes in places where it does
not add semantic value (because inheritance is a thing) we reference an
attribute we know exists.
This works because all our selectors are prefixed through
postcss-prefixwrap and "html" is (rightfully) detected as a "root tag".

[0] https://caniuse.com/#feat=css-dir-pseudo

Bug: T222695
Change-Id: I01a55edfda6cd6fc66f68b19591362edfc66f1cf