AnonEditWarning: checkbox to persist decision
Introduce a checkbox the value of which gets persisted as user
preference when either of the AnonEditWarning decisions is chosen.

Contains a change to Checkbox style to allow for labels to wrap -
there is no evidence which suggests that this is a feature that
may be undesired in other places (which don't even exist yet).

Bug: T223985
Depends-On: I1711de05efefdc3289c586b272951ae18df74490
Change-Id: I46ae5668a7a28a5b187ac6b60a9d9f2e0defb683