Alter buildOpenApiSpec to not modify the HealthCheckQuery
The validator was parsing and modifying the query. This is needed
in the main termbox query logic but we do want it unparsed in the
x-amples block.

i.e. we want the revision to remain a string and the
preferredLanguages not to be parsed to an array but remain a
pipe separated string.

Change-Id: I53fab0aca2f01cfbb966c453004192a18efc8f38