npm: update
This is (minor and patch) updates to
+ @types/express@4.17.0
+ tslib@1.10.0
+ fork-ts-checker-webpack-plugin@1.3.5
+ @types/node@10.14.9
+ @storybook/addons@5.1.3
+ @typescript-eslint/eslint-plugin@1.10.2
+ typescript@3.5.1
+ @storybook/addon-storysource@5.1.3
+ @storybook/cli@5.1.3
+ @storybook/vue@5.1.3

This also makes @typescript-eslint/parser a direct (dev) dependency
which apparently is only a peer dependency in other packages now.

Skips service-runner@2.7.1 as it was acting up
"node_modules/service-runner: Appears to be a git repo or submodule. Refusing to remove it."
afterwards

We are currently faced with a message
"Browserslist: caniuse-lite is outdated. Please run next command `npm update caniuse-lite browserslist`"
but running the stated command does not currently resolve it - waiting.

This leaves us with 65 vulnerabilities (64 low, 1 moderate)
* https://nodesecurity.io/advisories/577    1x
* https://nodesecurity.io/advisories/782    1x
* https://nodesecurity.io/advisories/786   63x

Manually verified that the blubber production variant works as expected.

Change-Id: Ie44908c875d12dc6781df381c3b036d4967579f2