termbox-updater: run every 15mins
Ensure to run the termbox-updater only every so often to ensure we don't
put too much strain on the github API. Hitting the rate limit could
create problems for our service and cause other trouble for the IPs we
are on.

I applied this to beta already to avoid problems.

See https://developer.github.com/v3/#rate-limiting

Bug: T225858
Change-Id: Ia54578d6215fa004d61203e7e4531d78da5d6867