Add context to errors to allow for more centralized logging
This introduces the LoggableError which should be implemented by all
exceptions that are logged in `app.ts`. For now this includes
`TechnicalError`s of any kind. In the future we also want this to be the
case for invalid requests and `BundleBoundaryPassingException`s which
get logged with a different log level.

By adding the `AxiosError` context to the exceptions thrown in the data
fetching repositories, `AxiosErrorLogger` became obsolete, thereby also
removing the redundant logging of request errors, which before happened
once through `errorLoggingInterceptor` and once through the
`TechnicalProblem` in app.ts.

Bug: T218081
Change-Id: I6402ab050f251653828611465fd01c2083b4e83e