Adjustments to eslintrc
During the termbox hike and now in the data-bridge hike was mentioned,
that it would benefical to get the eslint settings from a dedicated
place.
To make this happen this patch makes some initial aligments to the
existing config in termbox and will use the same config in data-bridge
as well. This will be a also the base for a upcomming pull request to https://github.com/wikimedia/eslint-config-wikimedia.
It also algines with
https://eslint.org/docs/user-guide/configuring#configuration-file-formats
in terms of not using deprecated formats anymore.
See also: I2df9b5e31d90546db458e9c8dfabb0d3601b2434
and: I22115b36ea4864ff98761c5ba9e842316b1e5329

Note: This pathc contains a workaround for the "no-undef" issue. For tracking or futher reading see: https://github.com/typescript-eslint/typescript-eslint/issues/131 and https://github.com/eslint/typescript-eslint-parser/issues/437

Bug: T225952
Change-Id: I6c635edb9c75cc79031bcda5d0c7619ba2218a4d