Re-add padding to externally styled `.external` link
Styles of the external link are partially reset via reset-css in
`App.vue`.
Whitelisting the `.external` class for a-elements would be the
more future-proof solution but has proven challenging to implement. This
is a trade-off that works for now. We looked at that in
I93a65d9d6a61539efce37243a1d92752331506be &
Ie1a4f266d4cba70658235378195b1e6b71e97922 but were not happy with either.

https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/skins/MinervaNeue/+/master/resources/skins.minerva.content.styles/links.less#51

Bug: T221835
Change-Id: I95e5f4034ec9a93e70962983bb26de028400c518