Document and simplify running production images locally
The command for running prod images locally specified in the README no
longer works for the latest images out of the box, as the server would
die when connecting to services (logstash and statsd) that do not exist
locally. Having a separate debug config that provides log output by 
default is likely going to be useful for debugging/testing locally.

Change-Id: Ia7b65fcd3b3d438e577631b5702a738c4d8e8f02