createMockableStore: make reusable
Rename createMockableStore to hotUpdateDeep and decouple it from termbox
so it can be used in other projects where it can be helpful.
My plan would be to create a dedicated package, e.g. "wmde/vuex-helpers"
and move hotUpdateDeep and store/util there.
Both can actually become src/ files because hot reloading of stores is
not strictly limited to test use.

Change-Id: I0016cb3c508c7a3920ab114bd45f4efd56de2121