eslint: use shared eslint config
Use the shared eslint config and fix violations.
Introduces some exceptions
* file-specific rules for some particular src files with tough nuts
* folder-wide no-explicit-any for store code for the action/getter
  trees
* folder-wide generic syntax which is helpful and low-risk in tests

This intentionally drags the dependency to eslint-config-wikimedia a
little because 0.15.0 introduced reportUnusedDisableDirectives as a
config variable which is incompatible with eslint 5 on which multiple of
our (indirect) dependencies still rely - it's quite a tree and I'd like
to update this independently.

Change-Id: I3e2e470a2dfd58c065c74a76f71feb748246f654