IndeterminateProgressBar: fine tune shadow
Adjust the progress bar "shadow" to align with how it looks like in
visual editor[0].
This, as first instructed in [1], intentionally breaks with
following native OOUI precisely[2] in order to achieve higher
consistency with what editors may be used from other places.
In the spirit of T220671, however, it refrains from using em for values
that are not font-related (and result in fractions of pixels).

[0] https://gerrit.wikimedia.org/g/mediawiki/extensions/VisualEditor/+/90bbfb56837a3b01b1238ec191af827dd0e0dde2/modules/ve-mw/preinit/styles/ve.init.mw.ProgressBarWidget-wikimediatheme.css#13
[1] https://phabricator.wikimedia.org/T222890#5368131
[2] https://doc.wikimedia.org/oojs-ui/master/demos/?page=widgets&theme=wikimediaui&direction=ltr&platform=desktop#Progress-bar-indeterminate

Also done in Iae339ab547c1f2cfc716b1ed4c61c6eacda72937 until we have a
shared component lib.

Bug: T232468
Change-Id: Iae339ab547c1f2cfc716b1ed4c61c6eacda72937