EntityInitializer: unlearn fix for poor serialization
Entities serialized in PHP used to contain an empty array where the data
model actually expects an object with dynamic (possibly no) properties.
This was long fixed during T138104, our work around can be removed once
it is safe to assume there are no installations with the old behavior
around.

Bug: T240674
Change-Id: I17b180fffb37004d6ace7ed03fa3566078b0544d