Selenium: Update to WebdriverIO v6
This also remodels the dependency of termbox to wdio-wikibase. termbox
contains browser tests which are implemented against a certain wdio
version. But those tests are not run as part of this repository but
instead as part of wikibase (see docs/tests.md). Consequently, stating
that termbox is dependent on wdio-wikibase is not quite correct (only
the specs are) and not really enforceable (wikibase does not install
termbox deps). To anyhow give developers a chance to know with which
webdriver version in mind to develop specs, the wdio-wikibase dep was
added as peer dependency instead. This also eliminates three
installation warnings which were shown beforehand (mwbot,
wdio-mediawiki, webdriverio not installed alongside wdio-wikibase).

Bug: T260942
Change-Id: I31ab41ca96ad60429c34d2b0d7425c70f5c1792c