Termbox: add test comparing Vue versions with MW
This test compares the Vue and Vuex versions listed
on our package.json with the Mediawiki versions of those.
Since we don't have Mediawiki access in CI this test uses
the npm package require-from-url in order to load the
vue file, so the version numbers may be extracted
and inspected.

Bug: T259468
Change-Id: I33dc3286e35a9b496c5c33dfabd9fc2df1344406