Tag edits with configuration from ResourceLoader
This adds support for tagging edits to AxiosWritingEntityRepository, and
makes the client entry point read the configuration for the tags (and
maybe other things in future) from a config.json file that is expected
to be require()-able after webpack has finished building the package,
just like for vue and vuex. This expects that Termbox will be loaded as
a ResourceLoader package files module [1] where config.json (in the same
directory as wikibase.termbox.main.js, i.e. in dist!) has contents
generated by ResourceLoader from the PHP configuration.

[1]: https://www.mediawiki.org/wiki/ResourceLoader/Package_files

Bug: T286775
Change-Id: I8f06e8d871bdd9b6e959d4c433f5ba5bd7dde8f0