Convert second batch of components to options API
All the ones that use the Messages mixin, but not any state functions.

Several prop declarations need to change, because apparently Vue doesn’t
have a concept of allowing props to be set to null explicitly: if it’s
required, it must be non-null; if you want to allow null, it must not be
required (and there needs to be a default value too). I’m not sure how
this worked before.

Bug: T296202
Change-Id: Iaa2379708b7bd40c189f2fbb665fc6c943a39804