Reintegrate wikibase-vuejs-components
They use vue-class-component, but we need pure-Vue versions for the
migration to Vue 3. The code is directly copied from Wikidata Bridge as
added in change I6f1a70ebf0 (Wikibase commit c73438713a), plus the fix
from change I9bf48ca2c7.

In the Edit components, :maxlength needs to change to :max-length so
that it actually refers to the maxLength prop of the component
(otherwise the tests fail because a warning about this is logged to the
console). I’m not 100% sure how this worked before.

Bug: T296202
Change-Id: I3112d38aa2a65ffedc930e30f961c05343741ecd