Simplify store initialization (asyncData in buildApp)
Previously, buildApp() supported calling an asyncData() method on any
component in the app, passing in the store and the termbox request.
However, in practice, this method was only defined on the root App
component. Let’s simplify this by moving that code into a separate
module and calling that one function explicitly. This makes buildApp()
much shorter, and the intention is that a follow-up change will more or
less remove that function entirely: we’ll need one version using
createMwApp() for the browser, and one version still using new App() for
SSR, at least until we migrate everything to Vue 3-only.

This also makes getChildComponents() unused, so we can remove that.
Otherwise, we would’ve had to update getComponentIdentifier() for Vue 3
at some point – the internal cid member exists in the migration build,
but not in pure Vue 3.

Bug: T296202
Change-Id: Idbf8090d8670b152814d682491fd8875a0428dd0