Split buildApp into buildAppMw and buildAppSsr
The previous buildApp() function is split into two: buildAppMw() is used
in the browser and uses MediaWiki’s createMwApp() function, which will
also be available after MediaWiki core migrates to Vue 3; buildAppSsr()
is used in server-side-rendering and still uses new App() for now. Once
we fully migrate to Vue 3, we can merge these into one function again;
in the meantime, I think the resulting code duplication is acceptable.

In the buildApp tests and in the dev entry point, we use an extremely
simplified createMwApp() mock implementation so that buildAppMw() works
outside of MediaWiki. This allows us to test both functions with the
same tests. (Fixes in other tests are trivial.)

Bug: T296202
Change-Id: Iee23225e45ae34278e71e99ac61ec7c3fa26a93d