Use Vue 3 at build time
I’m afraid this is a huge commit, but I see no way around it. At a very
high level, it combines all of the following steps:

- Use the Vue 3 migration build
- Update tests for @vue/test-utils v2 (VTU2)
- Update storybook to Vue 3
- Fix migration warnings from the migration build
- Stop using the migration build
- Update SSR to Vue 3

The SSR part does not work while the migration build is in place (the
vue → @vue/compat alias causes issues), and I don’t want to risk merging
a commit (and publishing a Docker image) without working SSR, so I think
this all has to happen together. (Some changes that *can* be separated
have already been split out, either before this change or after it.)

SSR still works in a reasonably similar way: `npm run build-server`
builds two directories, serverDist/ and serverBuild/; serverDist/
contains a “bundle” of the Vue app, and serverBuild/ contains a small
express app which uses a “bundle renderer” to respond to HTTP requests
with server-rendered HTML. In Vue 2, the infrastructure for this was
provided by Vue itself, but in Vue 3 that’s no longer the case, so we
use a third-party vue-bundle-renderer project (from the nuxt-contrib
GitHub organization) for the rendering, and a hand-written script
(make-bundle.js) to concatenate the bundle out of the individual built
files (replacing the previous plugins in vue.config.js). The motivation
for keeping it this way is that, if I understand it correctly, SSR in
Vite works more or less similarly, so I anticipate that this will make a
future migration from Vue CLI to Vite easier. (A possible alternative
would be to try directly using the @vue/server-renderer API in the
express app, without a bundle. But this would probably require a larger
build step for this app.) One important aspect is that we have to
externalize Vue for the bundle build now, otherwise the bundle and the
express app will have two different copies of Vue and the server-side
rendering won’t work (the renderer sets the “current rendering
instance”, a piece of Vue-internal global state).

The way that directives work in SSR also changes. There is no longer a
way to specify directives (operating on VNodes) in the SSR rendering
call, as far as I can tell; instead, we use “directive transforms”,
which are specified at build time (in the vue.config.js), and which
*generate code* to produce the actual props for an element. The focus
directive transform is of course trivial (no props); the inlanguage one
is a bit tricky, but not too bad. (The code it generates includes some
duplication, but I couldn’t find a way to extract auxiliary variables or
anything like that.) Note that it has to be written in JavaScript, not
TypeScript, since it’s directly require()d inside the vue.config.js.
(This also means the build can’t be parallel, since parallel building –
the default behavior – quietly discards non-serializable options.) The
unit test for the inlanguage directive (transform) is gone, since I
couldn’t think of a good way to unit-test the new version; note that
it’s integration-tested as part of the edge-to-edge tests, though.
(Added 2021-01-17: apparently SSR directives are now supposed to be
implemented using a getSSRProps function on the directives, but I
couldn’t get that to work; see change I6b1c2dfd13 for my attempt.)

In the tests, we upgrade to Jest 27 so that we can use the latest
version of the Vue Jest integration (@vue/vue3-jest), as well as VTU2.
Most of the changes aren’t worth mentioning in detail (there’s a lot of
busywork); one notable aspect is that the Termbox component tests change
from shallow to regular (deep) mounting. With shallow mounting, many of
these tests failed, and while some of them could be fixed by stubbing
more components with their real versions (like the tests already did to
some extent), this seems pointless to me: Termbox is the root component,
it’s okay to just test it deeply. Apart from that, some tests need the
new renderStubDefaultSlot option: VTU2 stubbed components no longer
render the contents of the default slot by default, but we rely on this
in some places.

In the SSR tests, we can’t use @vue/server-test-utils because no Vue 3
version of it is available yet, but fortunately @vue/server-renderer is
reasonably easy to use directly; for navigating the generated HTML, we
use the library that VTU uses under the hood, Cheerio (a kind of
server-side jQuery). Note that we need to mount() something
non-shallowly before we can use @vue/server-renderer without
interference, because VTU implements shallow mounting by setting an
internal global Vue hook which it never resets. (Instead of the fake
mount(), we could also call the internal transformVNodeArgs() API
directly, but that seems worse to me.)

TypeScript diagnostics for the tests are suppressed in the Jest config,
since there are some strange errors about @/ imports not being found
that I haven’t been able to figure out yet.

Bug: T296202
Change-Id: I588102c9b5f8ee28c330474019654d34ff0c769f
Co-Authored-By: Michael Große <michael.grosse@wikimedia.de>
Depends-On: I21641827780f4a7d8a5196c7dab14fbc3e4f4175