Update node-sass and sass-loader
The node-sass update bumps node-gyp to a version that supports Python 3;
the sass-loader update is needed for compatibility with that node-sass
version. (Eventually we should update to sass, formerly known as
dart-sass, but let’s keep this change smaller and stay on node-sass for
just a bit longer.)

No changes in build output.

Change-Id: Ie41c87cb3328b1a60d618939d555633472edb9b1