Use findComponent() with CSS selectors
This is required in VTU v2, but can be done before the big migration
already, as long as we update to the latest v1. Sadly, we temporarily
need some ts-ignore comments because the v1 type annotations don’t
include the needed overload (even though it works at runtime); we’ll be
able to remove that again in v2.

Change-Id: Ide2eccb776171f07596adb63684abbeffcaebb44