Prepare tests for Jest 27
setImmediate() is a nonstandard API supported mainly by Internet
Explorer and Node.js, so we should avoid using it in our tests. Using
setTimeout() instead seems to work fine (we just need to use an arrow
function to discard the second Promise argument).

Furthermore, because setImmediate() is not generally supported in
browsers, starting with Jest 27 the jsdom environment (which our
JestCustomEnvironment extends) actually removes this global (along with
clearImmediate()). However, the bundle-runner package, which we will use
in the edge-to-edge tests following the full migration to Vue 3, crashes
if these globals aren’t present (though I don’t think it actually uses
them – it just wants to sandbox them, or something). To avoid this
error, restore the globals in our JestCustomEnvironment, and do that
already before we actually migrate to Vue 3 and Jest 27 (it doesn’t hurt
to do it now).

Bug: T296202
Change-Id: Ib0256efa924d49838bc74bfe227ff210c4501bdc