Remove client/directives/index.ts
Instead of having this file modify the global Vue state as soon as it’s
imported, move the two relevant lines of code to the places where the
file was previously imported. Under Vue 3, this global API is gone, so
the code would have to change anyways; the index.ts could export some
function like extendVueEnvironment(), but to me that doesn’t seem worth
it just for two directives.

Bug: T296202
Change-Id: Ia11abbd7a38f66ecf441d8a8fb42b79e99fa378b