Avoid v-model in Checkbox story
This follows up change I7aa8d2e244 (commit 1621441343), where we avoided
v-model elsewhere in Termbox, because it doesn’t quite work properly –
apparently this problem persists even with non-migration Vue 3.

Note that propagating the change from the knob to the component works,
but from the component to the knob doesn’t – but this is also how it
works prior to this change, and might just be how knobs are supposed to
work, I’m not sure.

Bug: T296202
Change-Id: I63c8622a0e022a1320fa5f52955b0339e4c5a317