Load all of @storybook/addon-essentials, not just docs
This is recommended in general, but also, following the Vue 3 migration
(change I588102c9b5), addon-docs doesn’t seem to work on its own
anymore, but with addon-essentials it works. Whatever.

Note that the set of “essential” addons includes controls, the
designated replacement for knobs, but we still use knobs. We should
migrate that at some point, but for now it still works, the presence of
controls doesn’t break knobs as far as I can tell.

Bug: T296202
Change-Id: I1deb1a7be862f95759e8f72fc2e591b0a2c9320f