Update buildAndAttemptHydration for Vue 3
In Vue 3, unlike in Vue 2, hydration can’t really fail: if the DOM
doesn’t match, Vue logs a warning, updates it as needed, and then logs a
final console.error(). (The docs claim that it throws away the whole
DOM, but it actually tries to salvage as much of it as possible, as far
as I can tell.) This means we can remove our old discard-DOM-and-retry
code, which is nice.

However, looking at these warnings (which are available if you load the
dev version of Vue via ResourceLoader debug mode) shows that so far
there was a mismatch right at the root element: we were still trying to
call mount() with a selector for an element that we expected Vue to
*replace* with the app, whereas under Vue 3 the mount argument is
actually an element (selector) *below which* the app is mounted. In
WikibaseLexeme change Ic36fe5027a (commit f69a22284e), we worked around
this by mounting into a document fragment and then replacing the target
element with that fragment, but here we want to actually hydrate the
existing DOM, so we need a proper wrapper element to mount into. The
HTML sent by the server contains no suitable wrapper (the root element’s
parent is the div.wikibase-entityview-main, which contains other content
that we don’t want to throw away), so we create our own wrapper just
before calling mount(). (Before the wrapper, we insert an HTML comment
advising users against relying on this element. This idea is borrowed
from some skins that used to do the same thing – this was removed for
performance reasons in T281025, but here I think it’s justifiable.) We
check first if the wrapper already exists – that way, we can add it to
the server-side code later without having to worry about the JS code.
(This “Server-side code” will probably be the Wikibase PHP code, not
Termbox SSR.)

With the mount point fixed, some further mismatch warnings can be
resolved by changing some components to set their isServerRendered flag
in a mounted() hook instead of a beforeMount() hook: this way, during
mount, the virtual DOM will still match the server-rendered DOM, and
we’ll switch the component out of server-rendered mode only after
mounting. This requires some extra $nextTick() awaiting in tests.

Unfortunately, even with these mounted() hook fixes, there are still
some remaining hydration warnings, related to the Sectionedit component
and its custom / proprietary wb:sectionedit tag, which the Wikibase PHP
code either removes or completely empties. I think this will require
more effort to fix (T300176); in the meantime, we just hide the warnings
and error messages from Vue, since the mounting and hydration still
works well regardless.

Bug: T296202
Change-Id: Id0159a04356876faf779121014b64e3bbcbcd2b3