Compare current and latest image hash
This change updates the comparison made in
`restarter.sh` to only use the remote 'latest'
tag, since images are no longer tagged with commit
hashes.

Consequently, as we are attempting to pull the
latest docker image every 15 minutes, there is no
longer a need to seperate the restarter and updater
scripts, and thus `updater.sh` and associated
ansible commands are removed.

Bug: T304328
Change-Id: Id8dc140a4dea2d8a3101c335806904db342bfb46