Migrate to node16 and package-lock.json version 2
The main goal of this change was to work around the issues observed
when trying to upgrade to package-lock.json version 2 in T309176,
leading to the creation of T329553.
However, that caused some issues when CI runs
`npm ci --production-only && npm dedupe`.
This change now tries to tackle that issue by switching from the
very old node 12 to a newer version in order to be able to access a
more favorable module resolution options.

The smaller step of going to node14 did not work, because the docker
image docker-registry.wikimedia.org/nodejs14-devel does not exist
and the image docker-registry.wikimedia.org/nodejs14-slim has only
npm 6 (which is a major version down from npm 7 in
docker-registry.wikimedia.org/nodejs12-devel) and the version 2
format for package-lock.json was only introduced with npm 7.

This change also incorporates the changes from I4b8441ed and upgrades
eslint to v8 to tackle another peer dependency resolution issue.
node-sass is replaced with sass because the former seemed to lead to
error with node 16 and the change was planned anyway, too.

The changes to the dist/ files seem minor.

The version in .nvmrc is aligned with the node version in
docker-registry.wikimedia.org/nodejs16-slim. Both are using npm 8,
with the minor version differing slightly by default
(8.12.0 vs 8.15.0), which should hopefully not be an issue.

Bug: T328295
Bug: T329553
Change-Id: I1b6ea2f4f5a5defd86a4e31dc8fe8cdd140fb0b7