Update Jest to version 28
The package changes were done with:

    npm i -D {jest{,-environment-jsdom},@vue/vue3-jest,@types/jest,ts-jest}@28
    npm i @vue/test-utils@latest
    npm remove @types/lodash && npm i -S lodash-es && npm i -D @types/lodash-es

We have to set the customExportConditions to avoid @vue/test-utils
errors about Vue not being found (via [1]); however, this causes the
lodash/debounce import in ResizingTextField to fail in the tests. The
workaround from [2] to `import * as lodash` results in TypeScript errors
during `npm run build` (and build-server):

> TS2497: This module can only be referenced with ECMAScript
> imports/exports by turning on the 'allowSyntheticDefaultImports' flag
> and referencing its default export.

We can use lodash-es instead (suggested as the proper solution in [2]),
as long as we teach Jest to transform that package using babel-jest.

[1]: https://stackoverflow.com/a/72955178/1420237
[2]: https://github.com/thymikee/jest-preset-angular/issues/77#issuecomment-333965141

Bug: T309176
Change-Id: I72f0340de2d8baea580d8081c9c5e039f910e35b