Remove @vue/cli-plugin-unit-jest
This was still pulling in Jest 27, but we’re on Jest 28 now. (I’m not
even sure why change I72f0340de2 / commit 3985aa9774 worked at all, to
be honest.) It’s not too hard to replace, though.

The package changes were done with:

    npm r @vue/cli-plugin-unit-jest
    npm i -D jest-transform-stub
    npm i -D jest-serializer-vue

The environment variable in the test:unit command comes from [1]; for
some reason, it’s required (otherwise, lodash-es doesn’t get transpiled
and Jest complains about the `import` syntax inside it) only in CI –
even using Blubber, I was unable to reproduce the failures locally.

Also add babel-jest as an explicit dependency, since we’re referencing
it directly in jest.config.js; and stop using ts-jest as a preset: we
don’t need it and it might cause more confusion or issues.

[1]: https://github.com/vuejs/vue-cli/blob/f0f254e4bc81ed322eeb9f7de346e987e845068e/packages/%40vue/cli-plugin-unit-jest/index.js#L13-L14

Change-Id: I366d26886e8d065a5a0b8d1fcf7a71ebd3410d01